Add cli validation checks for backend-store-uri
